src/prefix_set: relocate inline tests to prefix_set_tests.rs
Same playbook as #1080-#1108. 242-LOC inline test block extracted.
After: prefix_set.rs drops from 505 → 266 LOC.
cargo build/test 92/878.
